+++ backend/auth-service/src/main/java/com/scopie/authservice/repository/MovieRepository.java
@@ -7,9 +7,9 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface MovieRepository extends JpaRepository<Movie, Integer> {
+public interface MovieRepository extends JpaRepository<Movie, Long> {
     @Query(value = "SELECT * FROM t_movie WHERE movie_id = ?1", nativeQuery = true)
-    Movie getMovieByMovieId(Integer movieId);
+    Movie getMovieByMovieId(Long movieId);
 
     @Query(value = "SELECT * FROM t_movie WHERE title LIKE '%?1%' OR genre LIKE '%?1%' OR language LIKE '%?1%'ORDER BY title", nativeQuery = true)
     List<Movie> findAllWithFilter(String filter);
