+++ backend/auth-service/src/main/java/com/scopie/authservice/service/ReservationService.java
@@ -6,11 +6,11 @@ import com.scopie.authservice.entity.Reservation;
 import java.util.Optional;
 
 public interface ReservationService {
-    public void reservationAcceptor(Integer reservationId, boolean allowance);
+    public void reservationAcceptor(Long reservationId, boolean allowance);
 
     void newReservation(ReservationDTO reservationDTO);
 
-    Optional<Reservation> getReservationById(Integer reservationId);
+    Optional<Reservation> getReservationById(Long reservationId);
 
-    void cancelReservation(Integer reservationId);
+    void cancelReservation(Long reservationId);
 }
